Make all splits into binary splits.
This commit removes the multi-way split support Brushfire has.
In practice splits end up being binary in most important cases,
and we end up with a lot of extra unnecessary predicates and
indirection.

As of this commit, all the tests pass and things more or less
work. There are some things that can still be cleaned up and
commented better, but this believed to be a moment where things
work.