Make Predicate[V] easier to work with.
In line with the move toward binary splits, we can simplify our
predicates to make them easier to work with. Previously we've relied a
lot on ad-hoc encodings, (such as Not(LessThan(_)) and
OneOf(LessThan(_), EqualTo(_)).

This commit replaces the previous predicate hierarchy with a set of
six non-recursive predicate types:

 - IsEq
 - NotEq
 - Lt
 - LtEq
 - Gt
 - GtEq

Predicates will display themselves in a more friendly format, which
should make them easier to work with. They also now come with a
negation operator (!p) which should remove the necessity of
creating/matching Not(_) nodes in user code.

The Ordering[V] requirement is moved from the Predicate[V]
constructors into the apply() and run() methods. This will reduce the
memory use for predicates, and also simplifies the serialization code.

The support for IsPresent(_) has also been removed. If you want
missing values to be treated specially (i.e. treated as just matching
or not matching), you'll need to arrange for that to happen before the
row value is constructed. Future work might generalize rows from
Map[K, V] to (K => V) to make this process more efficient.