This commit tightens up Brushfire's JSON injections.
Previously, we would accept a lot of values which
weren't precisely what Brushfire produced. In my
opinion this was potentially error-prone, since it
might cause future parsing probles that we woudn't
easily detect.

It also adds a test to ensure that our injections
will never throw exceptions on the inverse function,
even when fed arbitrary JSON.